move attribute to log field
